lib structure
